+++ release_notes.md
@@ -2,6 +2,7 @@
 
 ## NEXT
 * [TAB-37449] Show dual pricing and added processing fee on previewed order bill
+* [TAB-37740] Remove Paid in credit line from Credit slip documents (US only)
 
 ## 1.62.0
 * [TAB-37005] Add issuer country indication to the taxInvoice